Sprite updated
AnimSize int -> byte
currentFrameTimeLeft int -> ushort

+++ SonOfRobin.Core/Game/Animation/Sprite.cs
@@ -39,10 +39,10 @@ namespace SonOfRobin
         public LightEngine lightEngine;
         public ParticleEngine particleEngine;
         public AnimData.PkgName AnimPackage { get; private set; }
-        public int AnimSize { get; private set; }
+        public byte AnimSize { get; private set; }
         public string AnimName { get; private set; }
         private byte currentFrameIndex;
-        private int currentFrameTimeLeft; // measured in game frames
+        private ushort currentFrameTimeLeft; // measured in game frames
         public Rectangle GfxRect { get; private set; }
         public Rectangle ColRect { get; private set; }
 
@@ -54,7 +54,7 @@ namespace SonOfRobin
         public Cell currentCell; // current cell, that is containing the sprite
         public bool IsOnBoard { get; private set; }
 
-        public Sprite(World world, string id, BoardPiece boardPiece, AnimData.PkgName animPackage, int animSize, string animName, AllowedTerrain allowedTerrain, bool visible = true, LightEngine lightEngine = null)
+        public Sprite(World world, string id, BoardPiece boardPiece, AnimData.PkgName animPackage, byte animSize, string animName, AllowedTerrain allowedTerrain, bool visible = true, LightEngine lightEngine = null)
         {
             this.id = id; // duplicate from BoardPiece class
             this.boardPiece = boardPiece;
@@ -686,7 +686,7 @@ namespace SonOfRobin
         {
             if (this.AnimSize == newAnimSize) return;
 
-            int oldAnimSize = this.AnimSize;
+            byte oldAnimSize = this.AnimSize;
 
             if (setEvenIfMissing || CheckIfAnimSizeExists(newAnimSize))
             {
